+++ java-itu-minitwit/pom.xml
@@ -36,7 +36,7 @@
                         <manifest>
                             <addClasspath>true</addClasspath>
                             <classpathPrefix>lib/</classpathPrefix>
-                            <mainClass>minitwit</mainClass>
+                            <mainClass>Minitwit</mainClass>
                         </manifest>
                     </archive>
                 </configuration>
